% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_pois_noinfl.R
\name{E_pois_noinfl}
\alias{E_pois_noinfl}
\title{Expected Value for Positive Poisson Distribution}
\usage{
E_pois_noinfl(b, X)
}
\arguments{
\item{b}{Numeric vector of coefficients for the Poisson model.}

\item{X}{Matrix of predictors for the Poisson model, where rows correspond
to observations and columns to covariates.}
}
\value{
A numeric vector of expected values for each observation in the dataset.
}
\description{
Computes the expected value from a Poisson (PP) distribution based on the model parameters and covariates.
}
\examples{
\dontrun{
# Example inputs
b <- c(0.5, -0.2, 0.1)
X <- matrix(c(1, 2, 3, 1, 0.5, 1), ncol = 3)

# Compute expected values
E_pois_noinfl(b, X)
}

}
\seealso{
\code{\link{dEdq_pois_noinfl}} for computing partial derivatives of expected values in the Poisson model.
}
\keyword{internal}
