% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_negbin_noinfl.R
\name{E_negbin_noinfl}
\alias{E_negbin_noinfl}
\title{Expected Value for Zero-Truncated Negative Binomial Distribution}
\usage{
E_negbin_noinfl(b, a, X)
}
\arguments{
\item{b}{Numeric vector of coefficients for the regression model.}

\item{a}{Numeric scalar, the overdispersion parameter of the negative binomial distribution.}

\item{X}{Matrix of predictors for the model, where rows correspond to observations 
and columns to covariates.}
}
\value{
A numeric vector of expected values for each observation in the dataset.
}
\description{
Computes the expected value from a zero-truncated negative binomial (ZTNB) distribution 
based on the model parameters and covariates.
}
\seealso{
\code{\link{dEdq_nb_noinfl}} for computing partial derivatives of expected values in the ZTNB model.
}
\keyword{internal}
